requirements.txt
opencv-python
pytesseract
ahk
pillow